+++ example/responses/send-data.js
@@ -1,0 +1,9 @@
+module.exports = function (req, res) {
+    // Send success data to client.
+    return function (data) {
+        return this.json({
+            error: null,
+            data
+        });
+    };
+};